requirements.txt
torch>=1.9.0
torchvision>=0.10.0
scikit-learn>=0.24.0
opencv-python>=4.5.0
numpy>=1.21.0
pandas>=1.3.0
matplotlib>=3.4.0
seaborn>=0.11.0
flask>=2.0.0
Pillow>=8.3.0
scikit-image>=0.18.0
albumentations>=0.5.0
tqdm>=4.60.0
imutils>=0.5.0